+++ src/pagegen/constants.py
@@ -17,8 +17,6 @@ THEME_DIR = 'themes'
 THEME_TEMPLATE_DIR = 'templates'
 PLUGIN_DIR = 'plugins'
 LOCK_FILE = '.pgn_lock'
-LIVE_RELOAD_HASH_FILE = 'pagegen_site_hash'
-PGN_LIVE_RELOAD='PGN_LIVE_RELOAD'
 PAGEGEN_VERSION='4.0.0' # Managed by build.sh
 
 # Hooks