+++ CHANGELOG.md
@@ -7,12 +7,13 @@
 - Verbosity argument accepts level (DEBUG, INFO, WARNING, ERROR, CRITICAL)
 - log_level setting in site section of site.conf
 - Plugin site_search provides json index and javascript to implement a basic site search
+- Serve build dir locally managing base_url without changes to build dir. Replaces live reload
 
 ### Fixes
 
 - Create build dir if not exist
-- live reload clears cache ensuring base_url etc are right
 
 ### Removed
 
 - Excerpt plugin, suggest using headers instead
+- live reload